Redo,UndoをMain.tsxに移動
